+++ sopa/utils/__init__.py
@@ -2,6 +2,6 @@ from .image import (
     get_channel_names,
     scale_dtype,
     string_channel_names,
-    is_string_dtype,
+    valid_c_coords,
     _check_integer_dtype,
 )